+++ Changelog.md
@@ -3,10 +3,13 @@
 ## Added
 - Add a new character, the Thief. They will collect knowledge in the form of Dossiers
 - Add a handful of new upgrades for the thief
-- Added a new upgrade for the Apothecary, Secret Recipie
+- Added a new upgrade for the Apothecary, Perfect Potion
 - Added a new upgrade for the Culinarian, Ancient Cookbook
+- Added Plentiful Plundering upgrade for the Thief — awards gold per heist equal to dossiers × upgrade level
+- Added a new ranger upgrade, Fairy Hostage, that reveals the location of one randomly chosen pixie tile with a sparkle animation
 
 ## Removed
+- Dev Tools button... Unless?
 
 ## Changed